LayoutTests/http/tests/loading/slow-parsing-subframe-expected.txt

main frame - didStartProvisionalLoadForFrame
frame "f1" - willCloseFrame
main frame - willCloseFrame
main frame - didCommitLoadForFrame
main frame - didClearWindowObjectForFrame
frame "f1" - didClearWindowObjectForFrame
frame "f1" - didStartProvisionalLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "f1" - willCloseFrame
frame "f1" - didCommitLoadForFrame
frame "f1" - didClearWindowObjectForFrame
frame "f1" - didHandleOnloadEventsForFrame
main frame - didHandleOnloadEventsForFrame
frame "f1" - didFinishDocumentLoadForFrame
frame "f1" - didFinishLoadForFrame
main frame - didFinishLoadForFrame
This tests that we get frame load callbacks in the proper order when a subframe is parsed and created from within a timer callback.  
